Fix problem first and second stage.
Now the input is stored in an array of array of unsigned char so as to
read each fle per byte, this should allow me to read any type of file at
the same way.

Issue: I cannot able to print the size of the mtf output, strlen doesn't work properly even putting the terminator character.